Ref #63754
Add else condition to a diag2grad scripts
to prevent the script running when the
type/subtype case is not known.  Added
cases for uv subtypes 240,241,259,260,
286,290,291,296.

Modified time_vert.sh to preserve
stdout file from exec.

+++ util/Conventional_Monitor/nwprod/conmon_shared.v1.0.0/ush/diag2grad_ps_case.sh
@@ -25,6 +25,7 @@ set -xa
 
    echo "CONMON_NETCDF = ${CONMON_NETCDF}"
    netcdf=".false."
+   run_exe=1   
 
    if [ $CONMON_NETCDF -eq 1 ]; then
       netcdf=".true."
@@ -62,28 +63,34 @@ EOF
           run=${run},
 /
 EOF
+   else
+      run_exe=0
    fi
 
-   ./diag2grads <input>stdout 2>&1 
+   if [ $run_exe -eq 1 ]; then
+      ./diag2grads <input>stdout 2>&1 
 
-   rm -f ${mtype}_${subtype}.tmp
+      rm -f ${mtype}_${subtype}.tmp
 
 
-   mv stdout stdout_diag2grads_${mtype}_${subtype}.$run
-   dest_dir="${TANKDIR_conmon}/horz_hist/${run}"
+      mv stdout stdout_diag2grads_${mtype}_${subtype}.$run
+      dest_dir="${TANKDIR_conmon}/horz_hist/${run}"
 
-   grads_list=`ls ps*grads.${run}`
-   for file in $grads_list; do 
-      ${COMPRESS} ${file}
-      cp -f ${file}.${Z} ${dest_dir}/${file}.${PDATE}.${Z}
-   done
+      grads_list=`ls ps*grads.${run}`
+      for file in $grads_list; do 
+         ${COMPRESS} ${file}
+         cp -f ${file}.${Z} ${dest_dir}/${file}.${PDATE}.${Z}
+      done
 
-   scater_list=`ls ps*grads.${run}`
-   for file in $scater_list; do
-      ${COMPRESS} ${file}
-      cp -f ${file}.${Z} ${dest_dir}/${file}.${PDATE}.${Z}
-   done
+      scater_list=`ls ps*grads.${run}`
+      for file in $scater_list; do
+         ${COMPRESS} ${file}
+         cp -f ${file}.${Z} ${dest_dir}/${file}.${PDATE}.${Z}
+      done
 
+   else
+      echo "abort run, unmatched mtype ${mtype} " 
+   fi
 
    echo "<--  diag2grad_ps_case.sh"
